Internal function that augments ps and qs by filling in a grid of intermediate values for qs. n_grid new points are inserted between each pair of consecutive values in qs, and the corresponding ps are filled in by evaluating the spline output from spline_cdf_direct at the qs grid.

Internal function that augments ps and qs by filling in a grid of intermediate values for qs. n_grid new points are inserted between each pair of consecutive values in qs, and the corresponding ps are filled in by evaluating the spline output from spline_cdf_direct at the qs grid.

Usage

grid_augment_ps_qs(ps, qs, tail_dist, n_grid)